Pin pipenv to an older version (#88)
CI is currently failing because pipenv is seemingly unable to find dependencies from the rospypi index. I can reproduce this locally with pipenv versions 2023.7.23 and 2023.8.28. Locally, installing these dependencies directly with pip install --extra-index-url works fine, as does reverting to an older version of pipenv. Version 2023.6.18 was the version used in the latest successful CI run.

Prior versions of this PR attempted to remove the rosbag dependency and use an updated lockfile with latest pipenv, but that caused different errors, including std_messages to resolve to an incorrect version from pypi.org even though the index was specified.

This is a temporary workaround. We'll need to find a better fix here, but I don't have enough knowledge of this ecosystem to understand what's happening. I do see that version 2023.7.23 mentions changes related to compatibility checks.

The other changes here are minor:

- Upgrade CI setup actions to latest versions
- Specify python 3.8 as 3.7 is EOL