remove mutable default args (#136)
### Changelog

None

### Docs

None

### Description

In python defaulting to an empty object or array as an argument can lead
to unexpected behavior since the default is only evaluated once. See a
more detailed explanation here
https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments.

I don't believe this has been an issue because we do not modify the
arguments with a mutable default, nor do we return them from the
functions. Regardless, this should be fixed so we don't accidentally run
into trouble later.